backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
pytesseract==0.3.10
PyMuPDF==1.24.3
scikit-learn==1.4.2
httpx==0.27.0
